Fix data source configuration
`DataSourceProperties` is itself bound to the environment. If you are
using the same object with two completely different datasources, you
will not get the expected result.

Let's take a step back to better understand what happens by default in
Spring Boot. `DataSourceProperties` is used to share basic settings
about your datasource such as the url, the username, the password and
the datasource implementation to use. Additional settings can be set
using a specific namespace (i.e. `spring.datasource.hikari`) for Hikari
as not all implementations share the same settings.

There are several problems in your current setup.

First of all you are using the hikari namespace specific setup while you
are exposing your hikari datasource as `spring.datasource`. There's no
magic here: if you ask Spring Boot to expose your datasource as
`spring.datasource`, it won't read keys that start with
`spring.datasource.hikari`.

Second, which is the most problematic, you are using the same
`DataSourceProperties` for both datasources. It is exposed on
`spring.datasource` by default so it happens to work for the first
datasource only because you named it `spring.datasource` yourself. The
reason why it breaks is because you're trying to set the HSQL driver on
an H2 url. Hikari doesn't have a `setUrl` method so the `second.url`
attribute is ignored. You're therefore initializing the second datasource
with the basic settings of the first.

Before this commit, all basic settings from `spring.datasource` were set
on both datasource (since you're sharing the same `DataSourceProperties`)
and the `maximum-pool-size` parameter was ignored since you have a custom
namespace for your datasources.

I've reworked your configuration so that it does what is expected, I've
also renamed `spring.datasource` to `first` to make more obvious what is
happening. The first `DataSourceProperties` being primary, it will be the
one used by Spring Boot if it needs one (i.e. `DatabaseInitializer`). As
Spring Boot auto-detects the driver to use, I've commented that out since
it isn't necessary.