Add configuration properties metadata
This commit enables the Spring Boot configuration processor so that
custom metadata is generated at built time. Prior to this commit, the
`first` and `second` keys were displayed as unknown by the IDE.

In order to generate meta-data specific for Hikari, we now expose a more
dedicated type so that the annotation processor can pick that up. This is
exactly what we do internally to expose the standard
`spring.datasource.hikari` namespace.